Implement simple annotation system that simply inserts markers with magic values into each section. Markers exist for - start of a section, - end of a sections, - the desired position of the patchset, currently set to follow the end   of the section.
Basic preparation stages for the parser have been created. These
include an *annotation* stage and a *concat* stage.

The annotation stage simply adds the annotations discussed above.
The concat stage collects all the text from each section into a single
string value, and also builds an ordered collection for the patchsets.

Annotation is intended to happen before passing the sections into the
concat stage.

The patchsets are already in a kind of ast form, so they will be passed
to the renderer later on in order to nicely format them into web
elements.